pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[project]
authors = [
    # cSpell:disable
    { name = "Kohei Kumazawa", email = "48900989+dfc22@users.noreply.github.com" },
    { name = "Hodaka Ueda", email = "35979840+hodhodhod@users.noreply.github.com" },
    # cSpell:enable
]
dependencies = [
    "python-dotenv>=1.0.1",
    "faker>=28.4.1",
    "pillow>=10.4.0",
    "requests>=2.32.3",
    "numpy>=2.1.1",
    "torch>=2.4.1",
    "lightning>=2.4.0",
    "transformers>=4.44.2",
    "nltk>=3.9.1",
    "torchvision>=0.19.1",
]
description = "Add your description here"
name = "python-scaffold"
readme = "README.md"
requires-python = ">= 3.12"
version = "0.1.0"

[tool.hatch.build.targets.wheel]
packages = ["src", "script"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
line-length = 100

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "ANN1",  # missing type self and cls
    "E999",  # syntax error
    "UP027", # unpacked list comprehension
    "PT004", # pytest missing fixture name underscore
    "PT005", # pytest incorrect fixture name underscore
    "D1",    # undocumented
    "TD001", # invalid todo tag
    "TD002", # missing todo author
    "TD003", # missing todo link
]
unfixable = [
    "F401",   # unused import
    "F841",   # unused variable
    "ERA001", # commented out code
]

logger-objects = ["src.library.logger.LOGGER"]


[tool.rye]
dev-dependencies = [
    "ruff>=0.3.1",
    "pre-commit>=3.6.2",
    "types-requests>=2.32.0.20240907",
]
managed = true
# virtual = true


[tool.rye.workspace]
members = ["src", "scripts"]
useLibraryCodeForTypes = true
